requirements.txt
pandas
numpy
scikit-learn==1.6.1
joblib
gdown
google-generativeai
